Fixing ScaleFactor code after previous changes. Now compiles

+++ source/gbbCalibration/helpers/GlobalConfig.h
@@ -45,7 +45,7 @@ class GlobalConfig {
     return GetPtLabel(nonMuJetPt, m_NonMuJetPtBins, m_NonMuJetRegions);
   }
 
-  std::vector<float> GetBinning(const TString var);
+  std::vector<double> GetBinning(const TString var);
 
   TString GetMCHistName(const TString sys, const TString ptLabel, const TString flav, const TString var);
   TString GetDataHistName(const TString ptLabel, const TString var) {
@@ -67,7 +67,7 @@ class GlobalConfig {
   std::vector<TString> m_NonMuJetRegions;
 
   std::vector<TString> m_PlotVariables;
-  std::map<TString, std::vector<float> > m_PlotBinning;
+  std::map<TString, std::vector<double> > m_PlotBinning;
 
   std::vector<TString> MakeLabels(const std::vector<float> ptBins, const TString prefix);
   TString GetPtLabel(float pt, std::vector<float> ptBins, std::vector<TString> ptRegions);
